fix findbugs cloud jnlps
Someone reported that the following JNLP was not starting:
http://findbugs.cs.umd.edu/cloud/eclipse.jnlp

The first problem was that icedtea-web was trying to ensure that all
jnlps, including extension jnlps have a main class. I wonder what other
applications were broken by this check. This patch modifies the check
to only check for main class when a main class is possible (applets and
applications).

The second second problem was that (in the extension jnlp file) the
jnlp element did not contain a codebase attribute. IcedTea-Web tried to
add this null codebase when merging the jnlp file, causing an exception.
The patch simply ignores this null.

2012-01-05  Omair Majid <omajid@redhat.com>

       * netx/net/sourceforge/jnlp/runtime/JNLPClassLoader.java
       (initializeResources): Only throw exceptions about the main class not
       being found when the jnlp file can have a main class.
       (addToCodeBaseLoader): Dont try to process null URLs.