Fix caching of compressed content
To decide if an item is cached, icedtea-web does a HEAD request and
compares the Content-Length with the size of the cache entry. This
works fine when the jars are not compressed. When the jars are
fetched compressed and then uncompressed on disk, there is a
mismatch between the cache entry size and the remote Content-Length
and icedtea-web decides the content is not cached.

Store the original (compressed) size in the cache info file and use
that instead of the actual file size in the size-comparison

2014-05-12  Omair Majid  <omajid@redhat.com>

        * netx/net/sourceforge/jnlp/cache/CacheEntry.java: Add
        KEY_CONTENT_ORIGINAL_LENGTH, LENGTH_UNKNOWN.
        (getOriginalContentLength, setOriginalContentLength)
        (getLongKey(String,long)): New methods.
        (isCached): Check if the original content length is recorded and use it,
        if available, as the content length.
        * netx/net/sourceforge/jnlp/cache/ResourceTracker.java (downloadResource):
        If the content was compressed, store original content length in the cache
        entry.
        * tests/netx/unit/net/sourceforge/jnlp/cache/CacheEntryTest.java
        (testOriginalContentLengthIsSetCorrectly)
        (verifyCachedIfOriginalContentLengthsAreSame): New method.