fixed windows headless detection problems
* netx/net/sourceforge/jnlp/config/DeploymentConfiguration.java: as (mostly) first initialized class on windows, enforced headless check and initialized logging. This is WORKAROUND and shoudl get FIXED.
There is strange behavior in AWT windows impelmentation, which is causing awt queue to get corrupted when called from "bad" thread.  This hack is enforcing to be called from "correct" thread. This is  likely bug in JDK and have to be investigated. Btw see patch from 2017.04.10 about awt hooks. That is same issue.