ci: test against newer JDKs provided by Adopt too
ITW is intended to support Java 11 and up, and the people at
OpenWebStart are depending on that support to do their work as well. We
should at least test ITW against all versions of jdk that AdoptOpenJdk
is providing.

Since checkTopLevelWindow​ got removed we will only build on 8.

We enable all errors for the other two targets.