Add additional valid starting VM arguments
These were introduced in Java 8u191 (2018-10-16).

An important note on `MinRAMPercentage`: it is only used on ''small'' heap sizes
(think <100M). It is primarily used for container applications.

The properties match the following requirements:
* No `"` or `%` character (the double values do not include `%`)
* Values include `.` (0x2E) and 0-9 (0x30-0x39); that is between `0x20` and `0x17E`
* The strings do not include `\`

While these properties were originally introduced for containers, they are also
useful for desktop applications. Of specific note, `MaxRAMPercentage` is useful
for applications that can have a _small_ amount of data in memory to a _large_
amount of data in memory.

Default values:
* InitialRAMPercentage: 1.5625
* MinRamPercentage: 50.0
* MaxRamPercentage: 25.0